update home page
