+++ src/containers/sidebars/dashboard-sidebar/dashboard-home/dashboard-home-styles.module.scss
@@ -37,143 +37,41 @@
       color: var(--font-color);
     }
 
-    .explore{
-      position: relative;
+    .sections{
       display: flex;
-      justify-content: center;
-      align-items: center;
-      gap: 2rem;
-      margin-bottom: 20px;
+      gap: 20px;
+      flex-wrap: wrap;
+      justify-content: space-between;
 
-      .search{
-        border: 1px solid var(--border-color);
-        padding: 5px;
-        border-radius: 8px;
-        flex-grow: 1;
-        background: #{$white};
-
-        input[type="text"]{
-          color: var(--border-color);
-          width: 100%;
-        }
-
-        ::placeholder{
-          color: var(--border-color);
-        }
-
-        svg{
-          fill: var(--border-color);
-          color: var(--border-color);
-
-          path{
-            fill: var(--border-color);
-            color: var(--border-color);
-          }
-        }
-
-        &.showResults{
-          border-bottom-left-radius: 0;
-          border-bottom-right-radius: 0;
-        }
-      }
-
-      .searchResults{
-        position: absolute;
-        top: 100%;
-        left: 5px;
-        background-color: var(--border-color);
-        width: 329px;
-        border-bottom-right-radius: 5px;
-        border-bottom-left-radius: 5px;
-        padding: 8px 10px 5px;
-        margin: -2px 0 0 -1px;
-        list-style: none;
-
-        li{
-          button{
-            cursor: pointer;
-            padding: 5px;
-            color: #{$white};
-
-            span{
-              font-size: $font-size-sm;
-              font-style: italic;
-            }
-
-            &:hover{
-              background-color:#{$white};
-              color: var(--border-color);
-            }
-          }
-        }
-      }
-
-      button{
-        width:auto;
-        color: var(--save-button-text);
-        background-color: var(--border-color);
-        &:hover {
-          background-color: var(--border-color-hover);
-        }
-      }
-    }
-
-    .mostPopular{
-      .species{
+      .navCard{
         display: flex;
-        flex-wrap: wrap;
-        justify-content: center;
-        gap: 50px;
-        margin-top: 10px;
+        flex-direction: column;
+        color: var(--font-color);
+        fill: var(--font-color);
+        width: 250px;
 
-        .navCard{
+        .title{
           display: flex;
-          flex-direction: column;
-          border-radius: 15px;
-          height: 200px;
-          width: 200px;
-          cursor: pointer;
-          background-position: center top;
-          background-repeat: no-repeat;
-          background-size: cover;
-          padding: 0;
-          align-items: center;
-
-          img{
-            aspect-ratio: 4/3;
+          align-items: flex-end;
+          gap: 10px;
+          border-bottom: 1px solid var(--font-color);
+          width: 100%;
+          svg {
+            width: 40px;
+            height: 40px;
           }
 
-          span{
-            font-weight: 600;
-            font-size: $font-size-base;
-            letter-spacing: 0em;
-            display: inline-block;
-            padding-top: 5px;
-            text-align: center;
-            color: var(--dark-color);
-            background-color: var(--font-color);
-            margin: 0 12px;
-            transition: color .3s;
-            cursor: pointer;
-            width: 100%;
+          h2{
+            text-transform: uppercase;
+            margin: 0 0 5px;
+            font-size: $font-size-rg;
           }
+        }
 
-          p{
-            background-color: var(--font-color);
-            border-bottom-left-radius: 15px;
-            border-bottom-right-radius: 15px;
-            padding: 5px 15px 9px;
-            margin: 0;
-            text-align: center;
-            flex: 1;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            color: var(--dark-color);
-            font-size: $font-size-sm;
-            width: 100%;
-            font-style: italic;
-          }
+        p{
+          margin-top: 10px;
+          text-align: left;
+          font-size: $font-size-base;
         }
       }
     }
