+++ src/containers/sidebars/dashboard-sidebar/species-home/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import SpeciesHomeComponent from './species-home-component';
+
+function SpeciesHomeContainer(props) {
+  return <SpeciesHomeComponent {...props} />;
+}
+
+export default SpeciesHomeContainer;
