+++ src/containers/sidebars/dashboard-sidebar/species-home/species-home-styles.module.scss
@@ -1,0 +1,181 @@
+@import 'styles/ui.module';
+@import 'styles/settings';
+@import 'styles/typography-extends';
+@import 'styles/common-animations.module';
+
+.container {
+  --font-color: #{$white};
+  --dark-color: #{$black};
+
+  &.light{
+    --font-color: #{$black};
+    --dark-color: #{$white};
+  }
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 100%;
+  padding: 5px 10px;
+  height: max-content;
+  border-left: 1px solid var(--font-color);
+
+  .content{
+    flex-grow: 1;
+    height: 100%;
+    color: var(--font-color);
+
+    .sectionTitle{
+      font-size: $font-size-tick;
+      font-weight: 700;
+      text-transform: uppercase;
+      line-height: 1.5;
+      color: var(--font-color);
+    }
+
+    p{
+      color: var(--font-color);
+    }
+
+    .explore{
+      position: relative;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      gap: 2rem;
+      margin-bottom: 20px;
+
+      .search{
+        border: 1px solid var(--border-color);
+        padding: 5px;
+        border-radius: 8px;
+        flex-grow: 1;
+        background: #{$white};
+
+        input[type="text"]{
+          color: var(--border-color);
+          width: 100%;
+        }
+
+        ::placeholder{
+          color: var(--border-color);
+        }
+
+        svg{
+          fill: var(--border-color);
+          color: var(--border-color);
+
+          path{
+            fill: var(--border-color);
+            color: var(--border-color);
+          }
+        }
+
+        &.showResults{
+          border-bottom-left-radius: 0;
+          border-bottom-right-radius: 0;
+        }
+      }
+
+      .searchResults{
+        position: absolute;
+        top: 100%;
+        left: 5px;
+        background-color: var(--border-color);
+        width: 329px;
+        border-bottom-right-radius: 5px;
+        border-bottom-left-radius: 5px;
+        padding: 8px 10px 5px;
+        margin: -2px 0 0 -1px;
+        list-style: none;
+
+        li{
+          button{
+            cursor: pointer;
+            padding: 5px;
+            color: #{$white};
+
+            span{
+              font-size: $font-size-sm;
+              font-style: italic;
+            }
+
+            &:hover{
+              background-color:#{$white};
+              color: var(--border-color);
+            }
+          }
+        }
+      }
+
+      button{
+        width:auto;
+        color: var(--save-button-text);
+        background-color: var(--border-color);
+        &:hover {
+          background-color: var(--border-color-hover);
+        }
+      }
+    }
+
+    .mostPopular{
+      .species{
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: center;
+        gap: 50px;
+        margin-top: 10px;
+
+        .navCard{
+          display: flex;
+          flex-direction: column;
+          border-radius: 15px;
+          height: 200px;
+          width: 200px;
+          cursor: pointer;
+          background-position: center top;
+          background-repeat: no-repeat;
+          background-size: cover;
+          padding: 0;
+          align-items: center;
+
+          img{
+            aspect-ratio: 4/3;
+          }
+
+          span{
+            font-weight: 600;
+            font-size: $font-size-base;
+            letter-spacing: 0em;
+            display: inline-block;
+            padding-top: 5px;
+            text-align: center;
+            color: var(--dark-color);
+            background-color: var(--font-color);
+            margin: 0 12px;
+            transition: color .3s;
+            cursor: pointer;
+            width: 100%;
+          }
+
+          p{
+            background-color: var(--font-color);
+            border-bottom-left-radius: 15px;
+            border-bottom-right-radius: 15px;
+            padding: 5px 15px 9px;
+            margin: 0;
+            text-align: center;
+            flex: 1;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+            color: var(--dark-color);
+            font-size: $font-size-sm;
+            width: 100%;
+            font-style: italic;
+          }
+        }
+      }
+    }
+  }
+}
